requirements.txt
# ============================================
# EMBEDDING SERVICE REQUIREMENTS - FIXED
# ============================================
# Fixed LangChain compatibility with PaddleOCR

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.22

# Pydantic
pydantic==2.12.5
pydantic-settings==2.12.0

# Vector Database
qdrant-client==1.16.2

# Embedding Provider
openai==2.17.0

# Document Processing
PyMuPDF==1.26.7
Pillow==12.1.0

# OCR - Tesseract
pytesseract==0.3.13
opencv-python-headless>=4.8.0

# Text processing - FIXED for PaddleOCR compatibility
langchain==0.3.15
langchain-text-splitters==0.3.5
langchain-community==0.3.15
langchain-core==0.3.31 # Core dependencies

# Language Detection
langdetect==1.0.9

# Utilities
python-dotenv==1.2.1

# Development (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0